personal signature refactor
