+++ app/components/UI/PersonalSign/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PersonalSign';
